Pull out extract_containing_graph
